renamed range-slider
