+++ packages/@coorpacademy-components/demo/components.js
@@ -47,7 +47,7 @@ import ModuleBubble from '../src/molecule/module-bubble';
 import ModuleCard from '../src/molecule/module-card';
 import Pagination from '../src/molecule/pagination';
 import ProgressBar from '../src/molecule/progress-bar';
-import Range from '../src/molecule/range';
+import RangeSlider from '../src/molecule/range-slider';
 import ScopeContent from '../src/molecule/scope-content';
 import ScopeTabs from '../src/molecule/scope-tabs';
 import Search from '../src/molecule/search';
@@ -221,7 +221,7 @@ import PaginationFixtureDefault from '../src/molecule/pagination/test/fixtures/d
 import PaginationFixtureDisabled from '../src/molecule/pagination/test/fixtures/disabled';
 import ProgressBarFixtureDefault from '../src/molecule/progress-bar/test/fixtures/default';
 import ProgressBarFixtureMax from '../src/molecule/progress-bar/test/fixtures/max';
-import RangeFixtureDefault from '../src/molecule/range/test/fixtures/default';
+import RangeSliderFixtureDefault from '../src/molecule/range-slider/test/fixtures/default';
 import ScopeContentFixtureDefault from '../src/molecule/scope-content/test/fixtures/default';
 import ScopeContentFixtureEmpty from '../src/molecule/scope-content/test/fixtures/empty';
 import ScopeContentFixtureVideos from '../src/molecule/scope-content/test/fixtures/videos';
@@ -407,7 +407,7 @@ export const components = {
     ModuleCard,
     Pagination,
     ProgressBar,
-    Range,
+    RangeSlider,
     ScopeContent,
     ScopeTabs,
     Search,
@@ -695,8 +695,8 @@ export const fixtures = {
       Default: ProgressBarFixtureDefault,
       Max: ProgressBarFixtureMax
     },
-    Range: {
-      Default: RangeFixtureDefault
+    RangeSlider: {
+      Default: RangeSliderFixtureDefault
     },
     ScopeContent: {
       Default: ScopeContentFixtureDefault,
