+++ packages/@coorpacademy-components/src/molecule/range-slider/test/fixtures/default.js
