--- packages/@coorpacademy-components/src/molecule/range/style.css
@@ -1,14 +1,0 @@
-.default {
-}
-
-.rail {
-}
-
-.track {
-}
-
-.handle1 {
-}
-
-.handle2 {
-}
